consolidated demographic splits into a single plot (#37)
* consolidated demographic splits into a single plot

* cleaned up plot and moved more methods over to plotting.py

* Remove port mapping.
